+++ internal/models/loyalty.go
@@ -5,7 +5,7 @@ import "github.com/google/uuid"
 // Модель баланса лояльности
 type Loyalty struct {
 	UserID    uuid.UUID `json:"user_id"`
-	Balance   int       `json:"balance"`
+	Balance   float32   `json:"balance"`
 	CreatedAt string    `json:"created_at"`
 	UpdatedAt string    `json:"updated_at"`
 }
